Have kernel and X11 drivers use same size/type of 'wacom_drv' members
This fixes a bug where the pad buttons on the 21UX2 were not sending
X11 events. Kernel events appeared to be coming in without a problem,
but the X11 driver was being given bad values for 'npadkeys' and
'padkey_code'.

This bug is caused by the use of incorrect types for members of the
"wacom_drv" struct. It occurs when the Xorg driver is compiled as
32-bit and the kernel module as 64-bit. The kernel appears to be
accidentally using unsigned integers instead of the long time that
is assumed by the X11 driver. Switching the kernel to using longs
is insufficient for mixed 32/64 environments since it resutls in
the Xorg driver using 4-byte "longs" and the kernel 8-byte "longs".

This patch has both the kernel module and Xorg driver using the
uint32_t type which has a fixed size. Ideally all the "LONG" macros
should be changed to reflect the fact that they're actually
referencing the new type, but that is a farily trivial change.

+++ src/solaris/wacom.h
@@ -17,9 +17,9 @@ extern "C" {
 #define ERASER_DEVICE_ID	0x0A
 #define PAD_DEVICE_ID		0x0F
 
-#define BITS_PER_LONG		(sizeof(long)*8)
+#define BITS_PER_LONG		(sizeof(uint32_t)*8)
 #define BIT_WORD(n)		((n)/BITS_PER_LONG)
-#define BIT_MASK(n)		(((unsigned long)1) << ((n) % BITS_PER_LONG)) 
+#define BIT_MASK(n)		(((uint32_t)1) << ((n) % BITS_PER_LONG)) 
 
 enum {
 	PENPARTNER = 0,
@@ -64,11 +64,11 @@ struct wacom_drv {
         int tool[2];
         int id[2];
         uint32_t serial[2];
-	unsigned int keybit[NBITS(KEY_MAX)]; 
-	unsigned int evbit[NBITS(EV_MAX)];
-	unsigned int mscbit[NBITS(MSC_MAX)];
-	unsigned int relbit[NBITS(REL_MAX)];
-	unsigned int absbit[NBITS(ABS_MAX)];
+	uint32_t keybit[NBITS(KEY_MAX)]; 
+	uint32_t evbit[NBITS(EV_MAX)];
+	uint32_t mscbit[NBITS(MSC_MAX)];
+	uint32_t relbit[NBITS(REL_MAX)];
+	uint32_t absbit[NBITS(ABS_MAX)];
 	int abs_x;
 	int abs_y;
 	int abs_pressure;
